Pymc3 model adam (#48)
* getting started

I need help

* Error with model.py

changed to normal distribution instead of chi squared

* Fixed. Will message about changes here.

* Simplified some of the code layout.

* Added sampling to show how this works.

* created class for loading data

Not entirely sure if this is what you meant but here's my first attempt. It won't run on my computer due to an error with pymc3's handling of logspace in lines 99-101

* Fixed handling with class-based data loading. Redecorated with as_op

* added function to store data

* Run odeint in parallel.

* bump

* added build_model class

Should help with pickling. Not sure what to do with the plot on line 116.

* fixed plot