Handle distinct endosomal binding (#301)
This allows all the cytokines to have different binding kinetics in the endosome as compared to the surface. A function call endosomeAdjust(&endosome); can make the endosome have 5-fold weaker binding. This is left out of the full model parameterization so that master retains the same assumption as before, but can be added for Visterra work.

* weaker endosomal binding in full model
* can't use assignment operator to multiply by 5
* Move IL15 parameters to where they can be adjusted.
* Add valgrind.
* Moved over all params.
* Update model.hpp